chore: edit code for better readability
